fix : Router 로직 수정 및 add: Plan 페이지 관리 로직 추가
