+++ frontend/moaletter-app/src/pages/Plan/PlanDetails.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import Template from "../../components/Template";
+
+/**
+ * 모아모아의 상세 페이지
+ * @returns {JSX.Element} PlanDetails 페이지
+ */
+const PlanDetails = (): JSX.Element => {
+  return (
+    <Template>
+      <div className="my-6 mx-4">
+        <h1>상세페이지를 볼 수 있어요</h1>
+      </div>
+    </Template>
+  );
+};
+export default PlanDetails;
