--- frontend/moaletter-app/src/pages/PlanRegistration.tsx
@@ -1,17 +1,0 @@
-import React from "react";
-import Template from "../components/Template";
-
-/**
- * 모아모아를 등록하는 페이지
- * @returns {JSX.Element} PlanRegistration 페이지
- */
-const PlanRegistration = (): JSX.Element => {
-    return (
-        <Template>
-            <div className="my-6 mx-4">
-                <h1>모아모아를 등록해 보아요</h1>
-            </div>
-        </Template>
-    );
-};
-export default PlanRegistration;
